Edge cases in the DLL; MCM work
MCM now lets you set outfits as (in)active, and delete them. DLL now offers a function that can be used to force a visual update of an actor's equipment; the MCM calls this when closed to ensure that outfit changes take effect immediately. The DLL will no longer apply an outfit's shield to the player unless the player has any shield equipped. The DLL no longer overrides the player's equipment when no outfit is set as active (i.e. the player won't appear nude).