Backend fixes; frontend work
Added the ability to rename outfits. Added the ability to add armor to an outfit, choosing from what you currently have equipped. Tested the case of conflicting armors in an outfit (verdict: not allowed due to where and how we patch the engine), and added MCM-side code to prevent adding conflicting pieces to an outfit. Changed how ArmorAddonOverrideService keeps track of the current outfit, to fix issues that could occur when changes were made to the currently-active outfit (i.e. these changes wouldn't take effect before, because we copied the outfit when setting it as active).