Add to outfit, from list
You can now view a list of every armor defined in the game (and filter them by name or playable flag), and add any of those to an outfit. Templated armors are excluded from the list (i.e. you'll see "Ebony Armor" but not the dozen or so "Ebony Armor of EnchantmentName" variants).