Sorting
Outfit names are sorted alphabetically in the MCM. When adding armors from a list, the list will be sorted alphabetically. This needs some work -- I'd prefer to redesign our UTF8 code so that we can just use std::string without templating, and have case-insensitivity handled by the sort algorithm -- but for now, it works.