Critical fix for equipment conflicts
Fixed an issue where, when attempting to equip new armor, the game would run body slot conflict checks against the outfit instead of the armor. (Actor::Unk_120 is responsible for these checks; it runs them against the ActorWeightModel instead of the inventory for some reason.) We now replace the check in question.